Finalize dataset uploading
